requirements.txt
Django==3.0.1
asgiref==3.2.3
pytz==2019.3
djangorestframework==3.11.0